+++ src/_igraph/random.c
@@ -43,18 +43,14 @@ typedef struct {
   PyObject* rng_max_as_pyobject;
 } igraph_i_rng_Python_state_t;
 
-/* igraph_rng_get_int31() is potentially faster if the max value of the RNG
- * is 0x7FFFFFFF; however, in case of Python, it is actually _slower_ because
- * Python long integers are not terribly efficient. We are better off with using
- * any other value here */
-#define RNG_MAX 0xFFFFFFFF
-
-/* This must be consistent with the value of RNG_MAX above */
 #define RNG_BITS 32
+#define RNG_MAX ((((size_t) 1) << RNG_BITS) - 1)
 
-static igraph_i_rng_Python_state_t igraph_rng_Python_state = {0, 0, 0};
-static igraph_rng_t igraph_rng_Python = {0, 0, 0};
-static igraph_rng_t igraph_rng_default_saved = {0, 0, 0};
+static igraph_i_rng_Python_state_t igraph_rng_Python_state = {0};
+static igraph_rng_t igraph_rng_Python = {
+  /* type = */ 0, /* state = */ 0, /* is_seeded = */ 1
+};
+static igraph_rng_t igraph_rng_default_saved = {0};
 
 igraph_error_t igraph_rng_Python_init(void **state) {
   IGRAPH_ERROR("Python RNG error, unsupported function called",
@@ -111,7 +107,7 @@ PyObject* igraph_rng_Python_set_generator(PyObject* self, PyObject* object) {
   GET_FUNC("random"); new_state.random_func = func;
   GET_FUNC("gauss"); new_state.gauss_func = func;
 
-  /* construct the arguments of getrandbits(RNG_BITS) and randint(0, RNG_MAX)
+  /* construct the arguments of getrandbits(RNG_BITS) and randint(0, 2 ^ RNG_BITS-1)
    * in advance */
   new_state.rng_bits_as_pyobject = PyLong_FromLong(RNG_BITS);
   if (new_state.rng_bits_as_pyobject == 0) {
@@ -125,7 +121,7 @@ PyObject* igraph_rng_Python_set_generator(PyObject* self, PyObject* object) {
   if (new_state.one_as_pyobject == 0) {
     return 0;
   }
-  new_state.rng_max_as_pyobject = PyLong_FromUnsignedLong(RNG_MAX);
+  new_state.rng_max_as_pyobject = PyLong_FromSize_t(RNG_MAX);
   if (new_state.rng_max_as_pyobject == 0) {
     return 0;
   }
@@ -153,7 +149,7 @@ PyObject* igraph_rng_Python_set_generator(PyObject* self, PyObject* object) {
  * \ingroup python_interface_rng
  * \brief Sets the seed of the random generator.
  */
-igraph_error_t igraph_rng_Python_seed(void *state, unsigned long int seed) {
+igraph_error_t igraph_rng_Python_seed(void *state, igraph_uint_t seed) {
   IGRAPH_ERROR("Python RNG error, unsupported function called",
       IGRAPH_EINTERNAL);
   return IGRAPH_SUCCESS;
@@ -163,10 +159,10 @@ igraph_error_t igraph_rng_Python_seed(void *state, unsigned long int seed) {
  * \ingroup python_interface_rng
  * \brief Generates an unsigned long integer using the Python random number generator.
  */
-unsigned long int igraph_rng_Python_get(void *state) {
+igraph_uint_t igraph_rng_Python_get(void *state) {
   PyObject* result;
   PyObject* exc_type;
-  unsigned long int retval;
+  igraph_uint_t retval;
 
   if (igraph_rng_Python_state.getrandbits_func) {
     /* This is the preferred code path if the random module given by the user
@@ -198,7 +194,7 @@ unsigned long int igraph_rng_Python_get(void *state) {
       PyErr_WriteUnraisable(exc_type);
       PyErr_Clear();
     }
-    /* Fallback to the C random generator */
+    /* Fallback to the C random generator -- this should not happen */
     return rand() * RNG_MAX;
   } else {
     retval = PyLong_AsUnsignedLong(result);
@@ -226,7 +222,7 @@ igraph_real_t igraph_rng_Python_get_real(void *state) {
       PyErr_WriteUnraisable(exc_type);
       PyErr_Clear();
     }
-    /* Fallback to the C random generator */
+    /* Fallback to the C random generator -- this should not happen */
     return rand();
   } else {
     retval = PyFloat_AsDouble(result);
@@ -260,7 +256,7 @@ igraph_real_t igraph_rng_Python_get_norm(void *state) {
       PyErr_WriteUnraisable(exc_type);
       PyErr_Clear();
     }
-    /* Fallback to the C random generator */
+    /* Fallback -- this should not happen */
     return 0;
   } else {
     retval = PyFloat_AsDouble(result);
@@ -276,16 +272,19 @@ igraph_real_t igraph_rng_Python_get_norm(void *state) {
  */
 igraph_rng_type_t igraph_rngtype_Python = {
   /* name= */      "Python random generator",
-  /* min=  */      0,
-  /* max=  */      RNG_MAX,
+  /* bits= */      RNG_BITS,
   /* init= */      igraph_rng_Python_init,
   /* destroy= */   igraph_rng_Python_destroy,
   /* seed= */      igraph_rng_Python_seed,
   /* get= */       igraph_rng_Python_get,
-  /* get_real */   igraph_rng_Python_get_real,
+  /* get_int= */   0,
+  /* get_real= */  igraph_rng_Python_get_real,
   /* get_norm= */  igraph_rng_Python_get_norm,
   /* get_geom= */  0,
-  /* get_binom= */ 0
+  /* get_binom= */ 0,
+  /* get_exp= */   0,
+  /* get_gamma= */ 0,
+  /* get_pois= */  0,
 };
 
 void igraphmodule_init_rng(PyObject* igraph_module) {
@@ -295,8 +294,9 @@ void igraphmodule_init_rng(PyObject* igraph_module) {
     igraph_rng_default_saved = *igraph_rng_default();
   }
 
-  if (igraph_rng_Python.state != 0)
+  if (igraph_rng_Python.state != 0) {
     return;
+  }
 
   random_module = PyImport_ImportModule("random");
   if (random_module == 0) {